requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
alembic
psycopg2-binary
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
cloudinary
websockets
httpx
pytest
pytest-asyncio